+++ test/cypress/e2e/example05.cy.ts
@@ -1,4 +1,3 @@
-/// <reference types="cypress" />
 import { changeTimezone, zeroPadding } from '../plugins/utilities';
 
 describe('Example 05 - Tree Data (from a flat dataset with parentId references)', { retries: 1 }, () => {
@@ -183,8 +182,8 @@ describe('Example 05 - Tree Data (from a flat dataset with parentId references)'
           // only read the percent complete value if it's not a parent
           const $slickGroupToggleNotExpanded = $elm.children('.slick-cell:nth(0)').children('.slick-group-toggle:not(.expanded)');
           if ($slickGroupToggleNotExpanded.length > 1) {
-            const percentComplete = $elm.children('.slick-cell:nth(2)').first('div.percent-complete-bar-with-text').text();
-            expect(+percentComplete).to.be.lt(40)
+            const percentComplete = $elm.children('.slick-cell:nth(2)').first().text();
+            expect(+percentComplete).to.be.lt(40);
           }
         });
     }
@@ -236,7 +235,7 @@ describe('Example 05 - Tree Data (from a flat dataset with parentId references)'
     const today = changeTimezone(now, tz);
 
     const currentDate = today.getDate();
-    let currentMonth = today.getMonth() + 1; // month is zero based, let's add 1 to it
+    let currentMonth: number | string = today.getMonth() + 1; // month is zero based, let's add 1 to it
     if (currentMonth < 10) {
       currentMonth = `0${currentMonth}`; // add zero padding
     }